+++ flake.nix
@@ -5,7 +5,7 @@
     nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
     # Source only: meson builds seams-core as a static subproject.
     seams-core = {
-      url = "github:d-SEAMS/seams-core/98fb0b94b1c19d245b4494b6f43a982f9bbcf2cb";
+      url = "github:d-SEAMS/seams-core/7fbfc5148ca4da6c9ed063d951d7be19ad1dbf2a";
       flake = false;
     };
   };